# optimize cat performance on CPU with TensorIterator