@mingfeima 2019-12-05
Below shows the performance compare between original and this pr, the benchmark will compare from:

1. a range of input size: [64, 1000], [64, 10000], [128, 1000], [128, 10000]
2. size of the input array: 2, 16
3. contiguous tensor and non-contiguous tensor
4. single socket and single core

benchmark code available at this [op_bench-py](https://github.com/mingfeima/op_bench-py), to reproduce, `./run.sh cat.py`. Test machine is Xeon Skylake 6148 with 2*20 cores @2.40GHz

since `cat` is pretty much memory copy, the following table use the memory throughput as metric, unit: **GB/s**, the higher the better.

#### single socket run

input size | input num | contiguous | original (GB/s) | this pr (GB/s) | speedup
-- | -- | -- | -- | -- | --
[64, 1000] | 2 | TRUE | 15.586 | 38.35 | 2.46
[128, 1000] | 2 | TRUE | 9.586 | 70.78 | 7.38
[64, 10000] | 2 | TRUE | 6.076 | 303.51 | 49.95
[128, 10000] | 2 | TRUE | 2.692 | 7.29 | 2.71
[64, 1000] | 16 | TRUE | 12.344 | 22.15 | 1.79
[128, 1000] | 16 | TRUE | 4.089 | 5.33 | 1.30
[64, 10000] | 16 | TRUE | 2.645 | 4.87 | 1.84
[128, 10000] | 16 | TRUE | 2.634 | 7.28 | 2.76
[64, 1000] | 2 | FALSE | 25.986 | 35.93 | 1.38
[128, 1000] | 2 | FALSE | 34.263 | 63.92 | 1.87
[64, 10000] | 2 | FALSE | 46.875 | 277.85 | 5.93
[128, 10000] | 2 | FALSE | 6.843 | 7.84 | 1.15
[64, 1000] | 16 | FALSE | 22.539 | 21.85 | 0.97
[128, 1000] | 16 | FALSE | 3.755 | 5.75 | 1.53
[64, 10000] | 16 | FALSE | 4.68 | 4.85 | 1.04
[128, 10000] | 16 | FALSE | 7.083 | 7.37 | 1.04


#### single core run

input size | input num | contiguous | original (GB/s) | this pr (GB/s) | speedup
-- | -- | -- | -- | -- | --
[64, 1000] | 2 | TRUE | 16.321 | 16.05 | 0.98
[128, 1000] | 2 | TRUE | 7.073 | 12.33 | 1.74
[64, 10000] | 2 | TRUE | 6.339 | 12.65 | 2.00
[128, 10000] | 2 | TRUE | 2.683 | 4.25 | 1.59
[64, 1000] | 16 | TRUE | 11.986 | 12.77 | 1.07
[128, 1000] | 16 | TRUE | 3.589 | 4.69 | 1.31
[64, 10000] | 16 | TRUE | 2.655 | 3.04 | 1.14
[128, 10000] | 16 | TRUE | 2.63 | 3.14 | 1.19
[64, 1000] | 2 | FALSE | 3.28 | 14.56 | 4.44
[128, 1000] | 2 | FALSE | 3.313 | 11.91 | 3.59
[64, 10000] | 2 | FALSE | 3.412 | 12.90 | 3.78
[128, 10000] | 2 | FALSE | 2.144 | 4.13 | 1.93
[64, 1000] | 16 | FALSE | 3.347 | 12.53 | 3.74
[128, 1000] | 16 | FALSE | 2.251 | 4.55 | 2.02
[64, 10000] | 16 | FALSE | 1.913 | 3.03 | 1.58
[128, 10000] | 16 | FALSE | 1.929 | 3.16 | 1.64