@albanD 2019-12-06
Hi,

I'am curious about some lines in the first table above:
```
[64, 10000] | 2 | FALSE | 46.875 | 277.85 | 5.93
[128, 10000] | 2 | FALSE | 6.843 | 7.84 | 1.15
```

And more generally, all the results for this size see a large drop in speed transfer speed when we get to this size. Is this just an artefact of the caching in the processor? Or is there a deep meaning behind it?