@mingfeima 2019-12-23
> Hi,
> 
> I'am curious about some lines in the first table above:
> 
> ```
> [64, 10000] | 2 | FALSE | 46.875 | 277.85 | 5.93
> [128, 10000] | 2 | FALSE | 6.843 | 7.84 | 1.15
> ```
> 
> And more generally, all the results for this size see a large drop in speed transfer speed when we get to this size. Is this just an artefact of the caching in the processor? Or is there a deep meaning behind it?

@albanD  This is a very valuable question, yeah indeed `cache` seems to be a major difference here. But after some profiling with `vtune`, i find out that the real problem is `jemalloc` did out properly cache output tensor for the input size `[128, 10000]` with **DEFAULT** settings.

The thing is PyTorch operators usually generate output tensor at runtime and the memory allocation of large buffer may trigger **clear_page** which is part of memory management of linux kernel. And this could be a very big interference for CPU performance. So you may count on customized memory allocator such as `jemalloc`, `tcmalloc` or `tbbmalloc` to cache the memory so that subsequent `malloc` will be cached.

#### using `jemalloc` with default settings:
```bash
export LD_PRELOAD=/home/mingfeim/packages/jemalloc-5.2.0/lib/libjemalloc.so
### using OMP_NUM_THREADS=20
input size: [64, 10000]; input number: [2]; non-contiguous; thoughput: GB/s = 281.188
input size: [128, 10000]; input number: [2]; non-contiguous; thoughput: GB/s = 8.181
```
#### using `jemalloc` with customized settings:
```bash
export MALLOC_CONF="oversize_threshold:1,background_thread:true,metadata_thp:auto,dirty_decay_ms:9000000000,muzzy_decay_ms:9000000000";
export LD_PRELOAD=/home/mingfeim/packages/jemalloc-5.2.0/lib/libjemalloc.so
### using OMP_NUM_THREADS=20
input size: [64, 10000]; input number: [2]; non-contiguous; thoughput: GB/s = 284.893
input size: [128, 10000]; input number: [2]; non-contiguous; thoughput: GB/s = 146.153
```

The whole set of scripts is available at `run.sh` from [op_bench-py](https://github.com/mingfeima/op_bench-py)